requirements.txt
absl-py==0.6.1
alabaster==0.7.12
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
certifi==2019.11.28
chardet==3.0.4
cloudpickle==0.6.1
coverage==3.7.1
coveralls==1.1
cycler==0.10.0
Cython==0.29.2
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
docutils==0.14
editdistance==0.5.3
entrypoints==0.2.3
gast==0.2.1
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.17.1
h5py==2.9.0
idna==2.8
imageio==2.4.1
imagesize==1.1.0
imgaug==0.2.7
importlib-metadata==1.6.0
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.5.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.4.0
nbformat==4.4.0
nbstripout==0.3.3
networkx==2.2
nose==1.3.7
notebook==5.7.4
numpy==1.14.5
opencv-python==3.4.5.20
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pdf2image==1.5.1
pdfminer.six==20181108
Pebble==4.3.10
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.0
pycryptodome==3.8.1
Pygments==2.3.1
pyparsing==2.3.0
pyrsistent==0.16.0
pytest==4.4.1
pytest-cov==2.5.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
QtPy==1.9.0
requests==2.21.0
scikit-image==0.14.1
scipy==1.2.0
setuptools==39.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcontainers==2.1.0
Sphinx==1.8.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.1.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
TexSoup==0.2.0
timeout-decorator==0.4.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xmltodict==0.12.0
zipp==3.1.0
